PinnZoo.jl Documentation

Conventions

Quaternions use [$q_w$ $q_x$ $q_y$ $q_z$] order and represent body to world rotations

Floating base joints use [$x$ $y$ $z$ $q_w$ $q_x$ $q_y$ $q_z$]$ order where the position is in the world frame

Linear and angular velocities corresponding to a floating base joint are in the body frame

Inputs/Variables

nq $\rightarrow$ # of configuration variables (1 per revolute/prismatic joint, 7 for floating joint)

nv $\rightarrow$ # of velocity variables. Also the number of degrees of freedom (1 per revolute/prismatic joint, 6 for floating joint)

nx $\rightarrow$ # of states, nq + nv

nc $\rightarrow$ # of points on the robot that kinematics were generated for

q $\rightarrow$ configuration vector, size nq.

x $\rightarrow$ state vector, size nx. The first nq elements are the configuration and the last nv elements are velocity.

x_dot $\rightarrow$ state vector derivative, size nx. The first nq elements are the derivative of the configuration with respect to time, and the last nv are the acceleration (derivative of velocity).

tau $\rightarrow$ generalized force vector, size nv. Represents forces/torques at each degree of freedom.

For the dynamics, we assume the following manipulator equation and velocity kinematics:

\[M(x)\dot{v} + C(x) = \tau\]

\[\dot{q} = E(q)v\]

Velocity Kinematics

E(q) is typically the identity except in the following cases (not an exhaustive list):

When the configuration includes a quaternion, E(q) includes the mapping from angular velocity into a quaternion time derivative, respecting $\dot{q}q = 0$, the unit norm constraint at the velocity level. Because angular velocities are related to axis-angles, there is a factor of 2 that shows up, so the mapping back from $v$ to $\dot{q}$ which we refer to as $E\_T$ is not equal to $E^T$.

When the translation velocity of a floating is in the body frame, but the position is in the world frame, so E(q) includes a body-to-world rotation matrix.

Dynamics Functions

PinnZoo.M_func — Function

M_func(model::PinnZooModel, x::Vector{Float64})

Return the mass matrix of the model as a function of the configuration (x[1:model.nq])

PinnZoo.C_func — Function

C_func(model::PinnZooModel, x::Vector{Float64})

Return the dynamics bias of the model (coriolos + centrifugal + gravitational forces)

PinnZoo.forward_dynamics — Function

forward_dynamics(model::PinnZooModel, x::Vector{Float64}, τ::Vector{Float64})

Return the forward dynamics v̇ = M(x) \ (τ - C) where M is the mass matrix and C is the dynamics bias vector.

PinnZoo.forward_dynamics_deriv — Function

forward_dynamics_deriv(model::PinnZooModel, x::Vector{Float64}, τ::Vector{Float64})

Return a tuple of derivatives of the forward dynamics (v̇) with respect to x and τ

PinnZoo.inverse_dynamics — Function

inverse_dynamics(model::PinnZooModel, x::Vector{Float64}, v̇::Vector{Float64})

Return the inverse dynamics τ = M⩒ + C where M is the mass matrix and C is the dynamics bias vector

PinnZoo.inverse_dynamics_deriv — Function

inverse_dynamics_deriv(model::PinnZooModel, x::Vector{Float64}, v̇::Vector{Float64}

Return a tuple of derivatives of the inverse dynamics (τ) with respect to x and v̇

Kinematics Functions

PinnZoo.kinematics — Function

kinematics(model::PinnZooModel, x::Vector{Float64})

Return a list of the locations of each body in model.kinematics_bodies in the world frame.

PinnZoo.kinematics_jacobian — Function

kinematics_jacobian(model::PinnZooModel, x::Vector{Float64})

Return the jacobian of the kinematics function with respect to x (not projected into the tangent space).

PinnZoo.kinematics_velocity — Function

kinematics_velocity(model::PinnZooModel, x::Vector{Float64})

Return a list of the instantaneous linear velocities of each body in model.kinematics_bodies in the world frame.

PinnZoo.kinematics_velocity_jacobian — Function

kinematics_velocity_jacobian(model::PinnZooModel, x::Vector{Float64})

Return the jacobian of the kinematics_velocity function with respect to x (not in the tangent space). If $J_q$ is the derivative of the kinematics with respect to $q$, this jacobian $J$ = [$\dot{J_q}$ $J_q$] where $\dot{J_q} = \frac{\partial}{\partial q} J_qv$ and $J_qv$ is equal to kinematics_velocity. This also means that $J\dot{x}$ expresses the constraint at the acceleration level, i.e. $\dot{J_q}\dot{q} + J_q\dot{v} = 0$