PinnZoo.jl Documentation

PinnZoo.jl Documentation

Index

Dynamics Functions

Kinematics Functions

Utility Functions

Quaternion functions

Quadruped functions

Models

Conventions

Quaternions use [$q_w$ $q_x$ $q_y$ $q_z$] order and represent body to world rotations, using Hamilton's convention $i^2 = j^2 = k^2 = -1$

Floating base joints use [$x$ $y$ $z$ $q_w$ $q_x$ $q_y$ $q_z$]$ order where the position is in the world frame

Linear and angular velocities corresponding to a floating base joint are in the body frame

Inputs/Variables

nq $\rightarrow$ # of configuration variables (1 per revolute/prismatic joint, 7 for floating joint)

nv $\rightarrow$ # of velocity variables. Also the number of degrees of freedom (1 per revolute/prismatic joint, 6 for floating joint)

nx $\rightarrow$ # of states, nq + nv

nc $\rightarrow$ # of points on the robot that kinematics were generated for

q $\rightarrow$ configuration vector, size nq.

x $\rightarrow$ state vector, size nx. The first nq elements are the configuration and the last nv elements are velocity.

x_dot $\rightarrow$ state vector derivative, size nx. The first nq elements are the derivative of the configuration with respect to time, and the last nv are the acceleration (derivative of velocity).

tau $\rightarrow$ generalized force vector, size nv. Represents forces/torques at each degree of freedom.

For the dynamics, we assume the following manipulator equation and velocity kinematics:

\[M(x)\dot{v} + C(x) = \tau\]

\[\dot{q} = E(q)v\]

Velocity Kinematics

E(q) is typically the identity except in the following cases (not an exhaustive list):

When the configuration includes a quaternion, E(q) includes the mapping from angular velocity into a quaternion time derivative, respecting $\dot{q}q = 0$, the unit norm constraint at the velocity level. Because angular velocities are related to axis-angles, there is a factor of 2 that shows up, so the mapping back from $v$ to $\dot{q}$ which we refer to as $E\_T$ is not equal to $E^T$.

When the translation velocity of a floating is in the body frame, but the position is in the world frame, so E(q) includes a body-to-world rotation matrix.

Index

PinnZoo.Cartpole

PinnZoo.Go1

PinnZoo.Go2

PinnZoo.B_func

PinnZoo.C_func

PinnZoo.L_mult

PinnZoo.M_func

PinnZoo.R_mult

PinnZoo.apply_Δx

PinnZoo.attitude_jacobian

PinnZoo.axis_angle_to_quat

PinnZoo.error_jacobian

PinnZoo.error_jacobian_T

PinnZoo.fix_joint_limits

PinnZoo.forward_dynamics

PinnZoo.forward_dynamics_deriv

PinnZoo.init_state

PinnZoo.init_state

PinnZoo.init_state

PinnZoo.inverse_dynamics

PinnZoo.inverse_dynamics_deriv

PinnZoo.inverse_kinematics

PinnZoo.inverse_kinematics

PinnZoo.is_floating

PinnZoo.kinematics

PinnZoo.kinematics_jacobian

PinnZoo.kinematics_velocity

PinnZoo.kinematics_velocity_jacobian

PinnZoo.nearest_ik

PinnZoo.quat_conjugate

PinnZoo.quat_to_axis_angle

PinnZoo.quat_to_rot

PinnZoo.randn_state

PinnZoo.skew

PinnZoo.state_error

PinnZoo.zero_state

Dynamics Functions

PinnZoo.M_func — Function

M_func(model::PinnZooModel, x::Vector{Float64})

Return the mass matrix of the model as a function of the configuration (x[1:model.nq])

PinnZoo.C_func — Function

C_func(model::PinnZooModel, x::Vector{Float64})

Return the dynamics bias of the model (coriolos + centrifugal + gravitational forces)

PinnZoo.forward_dynamics — Function

forward_dynamics(model::PinnZooModel, x::Vector{Float64}, τ::Vector{Float64})

Return the forward dynamics v̇ = M(x) \ (τ - C) where M is the mass matrix and C is the dynamics bias vector.

PinnZoo.forward_dynamics_deriv — Function

forward_dynamics_deriv(model::PinnZooModel, x::Vector{Float64}, τ::Vector{Float64})

Return a tuple of derivatives of the forward dynamics (v̇) with respect to x and τ

PinnZoo.inverse_dynamics — Function

inverse_dynamics(model::PinnZooModel, x::Vector{Float64}, v̇::Vector{Float64})

Return the inverse dynamics τ = M⩒ + C where M is the mass matrix and C is the dynamics bias vector

PinnZoo.inverse_dynamics_deriv — Function

inverse_dynamics_deriv(model::PinnZooModel, x::Vector{Float64}, v̇::Vector{Float64}

Return a tuple of derivatives of the inverse dynamics (τ) with respect to x and v̇

Kinematics Functions

PinnZoo.kinematics — Function

kinematics(model::PinnZooModel, x::Vector{Float64})

Return a list of the locations of each body in model.kinematics_bodies in the world frame.

PinnZoo.kinematics_jacobian — Function

kinematics_jacobian(model::PinnZooModel, x::Vector{Float64})

Return the jacobian of the kinematics function with respect to x (not projected into the tangent space).

PinnZoo.kinematics_velocity — Function

kinematics_velocity(model::PinnZooModel, x::Vector{Float64})

Return a list of the instantaneous linear velocities of each body in model.kinematics_bodies in the world frame.

PinnZoo.kinematics_velocity_jacobian — Function

kinematics_velocity_jacobian(model::PinnZooModel, x::Vector{Float64})

Return the jacobian of the kinematics_velocity function with respect to x (not in the tangent space). If $J_q$ is the derivative of the kinematics with respect to $q$, this jacobian $J$ = [$\dot{J_q}$ $J_q$] where $\dot{J_q} = \frac{\partial}{\partial q} J_qv$ and $J_qv$ is equal to kinematics_velocity. This also means that $J\dot{x}$ expresses the constraint at the acceleration level, i.e. $\dot{J_q}\dot{q} + J_q\dot{v} = 0$

Utility Functions

PinnZoo.is_floating — Function

is_floating(model::PinnZooModel)

Return whether the model includes a floating base joint

PinnZoo.zero_state — Function

zero_state(model::PinnZooModel)

Return the neutral state of the model.

Warning

This currently defaults to all zeros unless overridden and will be wrong if there is a quaternion in the state.

PinnZoo.randn_state — Function

randn_state(model::PinnZooModel)

Return a state vector where every coordinate is drawn from a normal distribution

Warning

This currently does not respect the unit norm constraint on quaternions unless overridden

PinnZoo.init_state — Function

init_state(model::Go1)

Return state vector with the robot on the ground in the starting pose

init_state(model::Go2)

Return state vector with the robot on the ground in the starting pose

init_state(model::PinnZooModel)

Return a custom initial state for the model (like standing) if defined, otherwise returns all zeros

Warning

This currently defaults to all zeros unless overridden and will be wrong if there is a quaternion in the state.

Quaternion functions

PinnZoo.quat_to_axis_angle — Function

quat_to_axis_angle(q; tol = 1e-12)

Return the axis angle corresponding to the provided quaternion

PinnZoo.axis_angle_to_quat — Function

axis_angle_to_quat(ω; tol = 1e-12)

Return the quaternion corresponding to the provided axis angle

PinnZoo.quat_conjugate — Function

quat_conjugate(q)

Return the conjugate of the given quaternion (negates the velocity part)

PinnZoo.skew — Function

skew(v)

Return a matrix M such that $v \times x = Mx$ where $\times$ denotes the cross product

PinnZoo.L_mult — Function

L_mult(q)

Return a matrix representation of left quaternion multiplication, i.e. $q1 \cdot q2 = L(q1)q2$ where $\cdot$ is quaternion multiplication.

PinnZoo.R_mult — Function

R_mult(q)

Return a matrix representation of right quaternion multiplication, i.e. $q1 \cdot q2 = R(q2)q1$ where $\cdot$ is quaternion multiplication.

PinnZoo.attitude_jacobian — Function

attitude_jacobian(q)

Return the attitude jacobian G define as $\dot{q} = 0.5G\omega$, mapping angular velocity into quaternion time derivative.

PinnZoo.quat_to_rot — Function

quat_to_rot(q)

Return a rotation matrix R that q represents, defined by $\hat{p}^+ = q\hat{p}q^\dagger$ where $\hat{p}$ turns $p$ into a quaternion with zero scalar part and $p$ as the vector part, and $^\dagger$ is the quaternion conjugate.

Quadruped functions

TODO: Generalize state error and related functions to all models

PinnZoo.state_error — Function

state_error(quad::Quadruped, x, x0)

Return the state_error between x and x0, using axis-angles for quaternion error, and representing body position error in the body frame (matches with body velocity convention).

PinnZoo.apply_Δx — Function

apply_Δx(quad::Quadruped, x_k, Δx)

Return Δx added to x while respecting the configuration space/tangent space relationship (i.e. do quaternion multiplication, rotate Δbody pos from body frame to world frame)

PinnZoo.error_jacobian — Function

error_jacobian(quad::Quadruped, x)

Return the jacobian mapping Δx to x where Δx is in the tangent space (look at state_error for our choice of tangent space).

PinnZoo.error_jacobian_T — Function

error_jacobian_T(quad::Quadruped, x)

Return the jacobian mapping x to Δx where Δx is in the tangent space (look at state_error for our choice of tangent space).

PinnZoo.B_func — Function

B_func(quad::Quadruped)

Return the input jacobian mapping motor torques into joint torques

PinnZoo.fix_joint_limits — Function

fix_joint_limits(model::Quadruped, x; supress_error = false)

Return x with joint angles wrapped to 2pi to fit within joint limits if possible. If not and supress_error is false, this will error. Otherwise, this will return a clamped version.

PinnZoo.nearest_ik — Function

nearest_ik(model::Quadruped, q, foot_locs; obey_limits = true)

Return the ik solution for the given foot_locs and body orientation in q that is closest to the current joint angles in q (defined by minimum norm per joint).

Models

Cartpole

PinnZoo.Cartpole — Type

Cartpole() <: PinnZooModel

Return a Cartpole dynamics model

Unitree Go1

PinnZoo.Go1 — Type

Go1(; μ = 0.3) <: Quadruped

Return the Unitree Go1 dynamics and kinematics model

PinnZoo.init_state — Method

init_state(model::Go1)

Return state vector with the robot on the ground in the starting pose

PinnZoo.inverse_kinematics — Method

inverse_kinematics(model::Go1, x, foot_locs)

Return 12 by 2 matrix containing the 2 possible joint angle solutions that put the feet at foot_locs given the body pose in x. Will be populated with NaN or Inf if solutions don't exist.

Unitree Go2

PinnZoo.Go2 — Type

Go2(; μ = 0.3) <: Quadruped

Return the Unitree Go2 dynamics and kinematics model

PinnZoo.init_state — Method

init_state(model::Go2)

Return state vector with the robot on the ground in the starting pose

PinnZoo.inverse_kinematics — Method

inverse_kinematics(model::Go2, x, foot_locs)

Return 12 by 2 matrix containing the 2 possible joint angle solutions that put the feet at foot_locs given the body pose in x. Will be populated with NaN or Inf if solutions don't exist.

IHMC Nadia